+++ app/src/main/java/code/name/monkey/retromusic/adapter/song/SongAdapter.kt
@@ -177,7 +177,7 @@ open class SongAdapter(
                             imageContainerCard ?: image,
                             "${activity.getString(R.string.transition_album_art)}_${song.albumId}"
                         )
-                        NavigationUtil.goToAlbumOptions(activity, song.albumId, activityOptions)
+                        NavigationUtil.goToAlbumOptions(activity, song.albumId.toLong(), activityOptions)
                         return true
                     }
                 }
