+++ app/src/main/java/code/name/monkey/retromusic/util/MusicUtil.java
@@ -262,7 +262,7 @@ public class MusicUtil {
     }
 
     @NonNull
-    public static Uri getMediaStoreAlbumCoverUri(int albumId) {
+    public static Uri getMediaStoreAlbumCoverUri(long albumId) {
         final Uri sArtworkUri = Uri.parse("content://media/external/audio/albumart");
         return ContentUris.withAppendedId(sArtworkUri, albumId);
     }
@@ -322,7 +322,7 @@ public class MusicUtil {
         return songCount + " " + songString;
     }
 
-    public static Uri getSongFileUri(int songId) {
+    public static Uri getSongFileUri(long songId) {
         return ContentUris.withAppendedId(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI, songId);
     }
